+++ src/main/java/com/example/demo/repository/LibroRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface LibroRepository extends JpaRepository<Libro, Integer> {
+public interface LibroRepository extends JpaRepository<Libro, Long> {
 
     Optional<Libro> findByTitulo(String titulo);
     List<Libro> findByStockGreaterThan(Integer cantidad);
